+++ HISTORY.md
@@ -2,5 +2,11 @@
 
 ## Version 0.1.0
 
+- Release date: Dec 06, 2020
+- Db Connect module: added MongoDB Connection feature
+
+
+## Version 0.1.0
+
 - Release date: Dec 05, 2020
-- Db Connect module: for RDBMS => PostgresDB, MySQL, SQLite
+- Db Connect module: for RDBMS => PostgresDB, MySQL, SQLite...